ci: move scripts and variables to utils subdir
