+++ .github/workflows/android.yml
@@ -42,7 +42,7 @@ jobs:
       - name: Load variables from file
         uses: antifree/json-to-variables@v1.0.1
         with:
-          filename: .github/workflows/variables.json
+          filename: .github/workflows/utils/variables.json
 
       - name: Setup Go
         uses: actions/setup-go@v2
