+++ .github/workflows/ios.yml
@@ -43,7 +43,7 @@ jobs:
           #    all            pick all the runner configurations (in parallel)
           #    optimized      pick a dc4 runner if available or fallback on github one
           #
-          node .github/workflows/mac-runner-matrix-builder.js optimized "${{ secrets.SELFHOSTED_STATUS_CHECK_TOKEN }}"
+          node .github/workflows/utils/mac-runner-matrix-builder.js optimized "${{ secrets.SELFHOSTED_STATUS_CHECK_TOKEN }}"
 
   build:
     needs: mac_runner_matrix_builder
@@ -70,7 +70,7 @@ jobs:
       - name: Load variables from file
         uses: antifree/json-to-variables@v1.0.1
         with:
-          filename: .github/workflows/variables.json
+          filename: .github/workflows/utils/variables.json
 
       - name: Post-checkout cleanup
         if: ${{ matrix.selfhosted }}
